# docs/hrpm-v1.openapi.yaml
openapi: 3.1.0
info:
  title: HRPM API
  version: 1.0.0
  description: API contract. Resource-specific fields and state rules are defined in 04-接口设计.md.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /workflow/tasks:
    get:
      operationId: listWorkflowTasks
      parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PageSize'}]
      responses: {'200': {$ref: '#/components/responses/PageSuccess'}}
...
  /workflow/tasks/{id}/approve:
    post:
      operationId: approveWorkflowTask
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: int64}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/WorkflowActionCommand'}}}}
      responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Error'}, '422': {$ref: '#/components/responses/Error'}}
  /workflow/tasks/{id}/reject:
    post:
      operationId: rejectWorkflowTask
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: int64}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/WorkflowActionCommand'}}}}
      responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Error'}, '422': {$ref: '#/components/responses/Error'}}
  /workflow/tasks/{id}/return:
    post:
      operationId: returnWorkflowTask
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: int64}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/WorkflowActionCommand'}}}}
      responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Error'}, '422': {$ref: '#/components/responses/Error'}}
  /workflow/tasks/{id}/transfer:
    post:
      operationId: transferWorkflowTask
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: int64}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/WorkflowActionCommand'}}}}
      responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}, '422': {$ref: '#/components/responses/Error'}}
  /workflow/tasks/instances/{id}:
    get:
      operationId: getWorkflowInstance
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: int64}}]
      responses: {'200': {$ref: '#/components/responses/Success'}, '404': {$ref: '#/components/responses/Error'}}
  /workflow/tasks/instances/{id}/withdraw:
    post:
      operationId: withdrawWorkflowInstance
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: int64}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/WorkflowInstanceActionCommand'}}}}
      responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Error'}, '422': {$ref: '#/components/responses/Error'}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
...
  responses:
    Success: {description: Successful response, content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}}
    PageSuccess: {description: Successful paged response, content: {application/json: {schema: {$ref: '#/components/schemas/PageResponse'}}}}
    Error: {description: Error response, content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}}
  schemas:
    ApiResponse:
      type: object
      required: [code, message, traceId]
      properties: {code: {type: string}, message: {type: string}, data: {}, traceId: {type: string}}
    PageResponse:
      allOf:
        - {$ref: '#/components/schemas/ApiResponse'}
        - {type: object, properties: {data: {type: object, required: [records, total, page, pageSize], properties: {records: {type: array, items: {}}, total: {type: integer}, page: {type: integer}, pageSize: {type: integer}}}}}
...
    WorkflowActionCommand: {type: object, required: [version, comment], properties: {version: {type: integer, minimum: 0}, comment: {type: string, minLength: 1}, transferToUserId: {type: integer, format: int64}}}
    WorkflowInstanceActionCommand: {type: object, required: [version, comment], properties: {version: {type: integer, minimum: 0}, comment: {type: string, minLength: 1}}}
...